Module special_functions::approximations

Approximations

Modules

linear
Linear interpolation

Functions

chebyshev
Evaluates a series of Chebyshev functions at x.

piecewise_chebyshev
Evaluates an arbitrary piecewise Chebyshev function at a particular point.

piecewise_polynomial
Evaluates an arbitrary piecewise single-variable polynomial at a particular point.

piecewise_polynomial_ratio
Evaluates an arbitrary piecewise ratio of single-variable polynomial at a particular point.

polynomial
Evaluates an arbitrary single-variable polynomial at a particular point.

polynomial_ratio
Evaluates an arbitrary ratio of single-variable polynomials at a particular point. The coefficient of the polynomial in the numerator are given in a, and b for the denominator.