Function special_functions::approximations::linear::linear

pub fn linear(data: &[(f64, f64)], x: f64) -> f64

Perform linear interpolation on data.

The data should be provided in an array of the shape [(x0, y0), (x1, y1), ...]. The interpolation is done linearly such that a point half-way in between x0 and x1 will yield a value half-way between y0 and y1.

Warning

The slice if input values must be sorted in strictly ascending x values. Any repeated values, unordered entries or NaN values will result in undefined behaviour.

Extrapolation

This linear interpolation repeats the boundary value for all values outside the domain of the interpolation data.