Function special_functions::particle_physics::kallen_lambda_sqrt

#[must_use]pub fn kallen_lambda_sqrt(a: f64, b: f64, c: f64) -> f64

Square root of the Kallen lambda function:

\lambda^{\frac{1}{2}}(a, b, c) = \sqrt{a^2 + b^2 + c^2 - 2ab - 2ac - 2bc}

This implementation is more precise than taking the square root of kallen_lambda in cases where the arguments span several orders of magnitude.

Example

use special_functions::particle_physics::{kallen_lambda, kallen_lambda_sqrt};

assert!((kallen_lambda_sqrt(5.0, 2.0, 0.5) - 1.5).abs() < 1e-14);
assert!((kallen_lambda(5.0, 2.0, 0.5).sqrt() - kallen_lambda_sqrt(5.0, 2.0, 0.5)).abs() < 1e-14);
assert!((kallen_lambda_sqrt(1.0, 1.0, 1.0) - 3f64.sqrt()).abs() < 1e-14);

Warning

The result returns the root of the absolute value, thus returning a result even if kallen_lambda is negative. It is up to the user to determine whether this is valid or not.