Module special_functions::particle_physics::pave_absorptive

Passarino-Veltman coefficient functions.

Calculation of the imaginary component of Passarino-Veltman coefficient functions. The code is generated algorithmically from the Mathematica package Package-X (arXiv:1503.01469).

As the general reduction algorithm is not implemented, only the following functions are implemented:

$\boldsymbol{A}_{\underbrace{0\dots0}_{2r}}$ for all values of $r$;

$\boldsymbol{B}_{\underbrace{0\dots0}_{2r}\underbrace{1\dots1}_{n_1}}$ for all values of $r$ and $n_1$;

$\boldsymbol{C}_{\underbrace{0\dots0}_{2r}\underbrace{1\dots1}_{n_1}\underbrace{2\dots2}_{n_2}}$ for all combinations of $r$, $n_1$ and $n_2$ in $\{0, 1, 2\}$;

$\boldsymbol{D}_{\underbrace{0\dots0}_{2r}\underbrace{1\dots1}_{n_1}\underbrace{2\dots2}_{n_2}\underbrace{3\dots3}_{n_3}}$ for the scalar case and for one of $r$, $n_1$, $n_2$ or $n_3$ being 1.

In all cases, the higher order functions are generally less accurate. Furthermore, parameter space which can lead to numerical instabilities are not accounted explicitly handled at this stage.

Functions

a

Absorptive part of the Passarin-Veltman coefficient function $\boldsymbol{A}$.

b

Absorptive part of the Passarin-Veltman coefficient function $\boldsymbol{B}$.

c

Absorptive part of the Passarin-Veltman coefficient function $\boldsymbol{C}$.

d

Absorptive part of the Passarin-Veltman coefficient function $\boldsymbol{C}$.