Macro special_functions::approx_fn

macro_rules! approx_fn {
    (
        $(#[$outer:meta])*
        $vis:vis fn $fn:ident(mod = $mod:ident, type = polynomial, outer = $o:path, inner = $i:path);
    ) => { ... };
    (
        $(#[$outer:meta])*
        $vis:vis fn $fn:ident(mod = $mod:ident, type = ratio, outer = $o:path, inner = $i:path);
    ) => { ... };
    (
        $(#[$outer:meta])*
        $vis:vis fn $fn:ident(mod = $mod:ident, type = chebyshev, outer = $o:path, inner = $i:path);
    ) => { ... };
}

Create a function from a module containing all the information for the split.

The module must contain the following:

fn lower(x: f64) -> f64 which provides a (generally Taylor series) approximation near the lower bound of the domain. This function is used for values of x smaller than the first split.

fn upper(x: f64) -> f64 which provides a (generally Taylor series) approximation near the upper bound of the domain. This functinon is used for values of x larger than the last split.

const COEFFICIENTS: &[...] contains a list of list of coefficients, as used in the relevant piecewise function.

const SPLITS: &[f64] contains a list of splits. Within the splits, the relevant piecewise function is used.

The macro is used as:

approx_fn! {
    fn foo(mod = path::to::module, type = approx_type, outer = fn, inner = fn);
}

The approx_type can be either one of polynomial, polynomial_ratio or chebyshev. The outer and inner are names of functions which apply to the result and argument before they go in the piecewise approximation. These do not apply to lower and upper.