Module special_functions::particle_physics::statistics

Particle Statistics

Compute Bose-Einstein and Fermi-Dirac integrals, in both massive ($\mu = 0$) and massless ($\mu \inR$) cases as well as normalized to a single massless bosonic degree of freedom for the massive case.

Functions

bose_einstein_massive

Equilibrium number density of massive Bose-Einstein particle.

bose_einstein_massless

Equilibrium number density of a massless Bose-Einstein particle with chemical potential $\mu \inR$.

bose_einstein_normalized_massive

Equilibrium number density of massive Bose-Einstein particle normalized to a massless Bose-Einstein particle.

bose_einstein_normalized_massless

Equilibrium number density of massless Bose-Einstein particle with chemical potential $\mu \inR$ normalized to a massless Bose-Einstein particle.

fermi_dirac_massive

Equilibrium number density of massive Fermi-Dirac particle.

fermi_dirac_massless

Equilibrium number density of a massless Fermi-Dirac particle with chemical potential $\mu \inR$.

fermi_dirac_normalized_massive

Equilibrium number density of massive Fermi-Dirac particle normalized to a massless Bose-Einstein particle.

fermi_dirac_normalized_massless

Equilibrium number density of massless Fermi-Dirac particle with chemical potential $\mu \inR$ normalized to a massless Bose-Einstein particle.